refactor: [Session2/Step1] 문자열 덧셈 계산기 #6 (리팩토링)
패키지 변경
